// client/stocklinx app/src/redux/product/constant.ts
export enum productConst {
  FETCH_ENTITY_COUNTS_REQUEST = "FETCH_ENTITY_COUNTS_REQUEST",
  FETCH_ENTITY_COUNTS_SUCCESS = "FETCH_ENTITY_COUNTS_SUCCESS",
  FETCH_ENTITY_COUNTS_FAILURE = "FETCH_ENTITY_COUNTS_FAILURE",
  FETCH_PRODUCT_STATUS_COUNTS_REQUEST = "FETCH_PRODUCT_STATUS_COUNTS_REQUEST",
  FETCH_PRODUCT_STATUS_COUNTS_SUCCESS = "FETCH_PRODUCT_STATUS_COUNTS_SUCCESS",
  FETCH_PRODUCT_STATUS_COUNTS_FAILURE = "FETCH_PRODUCT_STATUS_COUNTS_FAILURE",
  FETCH_PRODUCT_LOCATION_COUNTS_REQUEST = "FETCH_PRODUCT_LOCATION_COUNTS_REQUEST",
  FETCH_PRODUCT_LOCATION_COUNTS_SUCCESS = "FETCH_PRODUCT_LOCATION_COUNTS_SUCCESS",
  FETCH_PRODUCT_LOCATION_COUNTS_FAILURE = "FETCH_PRODUCT_LOCATION_COUNTS_FAILURE",
  FETCH_PRODUCT_CATEGORY_COUNTS_REQUEST = "FETCH_PRODUCT_CATEGORY_COUNTS_REQUEST",
  FETCH_PRODUCT_CATEGORY_COUNTS_SUCCESS = "FETCH_PRODUCT_CATEGORY_COUNTS_SUCCESS",
  FETCH_PRODUCT_CATEGORY_COUNTS_FAILURE = "FETCH_PRODUCT_CATEGORY_COUNTS_FAILURE",
  FETCH_CUSTOM_LOGS_REQUEST = "FETCH_CUSTOM_LOGS_REQUEST",
  FETCH_CUSTOM_LOGS_SUCCESS = "FETCH_CUSTOM_LOGS_SUCCESS",
  FETCH_CUSTOM_LOGS_FAILURE = "FETCH_CUSTOM_LOGS_FAILURE",
}